+++ src/test/java/techcourse/myblog/controller/HomeControllerTests.java
@@ -3,6 +3,7 @@ package techcourse.myblog.controller;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
+import techcourse.myblog.controller.template.ControllerTestTemplate;
 
 public class HomeControllerTests extends ControllerTestTemplate {
     @Test
