+++ src/test/java/techcourse/myblog/controller/template/ControllerTestTemplate.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.controller;
+package techcourse.myblog.controller.template;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
